+++ src/resources/views/fields/table.blade.php
@@ -1,16 +1,22 @@
 <!-- array input -->
 
-@php
+<?php
     $max = isset($field['max']) ? $field['max'] : -1;
     $min = isset($field['min']) ? $field['min'] : -1;
     $item_name = strtolower( isset($field['entity_singular']) && !empty($field['entity_singular']) ? $field['entity_singular'] : $field['label']);
+
     $items = old($field['name']) ? (old($field['name'])) : (isset($field['value']) ? ($field['value']) : (isset($field['default']) ? ($field['default']) : '' ));
-    if( empty($items) ){
+
+    if(!is_array($items)) {
+        $items = json_decode($items);
+    }
+
+    if (empty($items)) {
         $items = '[]';
-    } elseif( is_string($items) && !is_array(json_decode($items)) ){
+    } elseif ( is_string($items) && !is_array(json_decode($items)) ) {
         $items = '[]';
     }
-@endphp
+?>
 <div ng-app="backpackTable" ng-controller="tableController" @include('crud::inc.field_wrapper_attributes') >
 
     <label>{!! $field['label'] !!}</label>
@@ -19,7 +25,7 @@
 
     <div class="array-container form-group">
 
-        <table class="table table-bordered table-striped m-b-0" ng-init="field = '#{{ $field['name'] }}'; items = {{$items}}; max = {{$max}}; min = {{$min}}; maxErrorTitle = '{{trans('backpack::crud.table_cant_add', ['entity' => $item_name])}}'; maxErrorMessage = '{{trans('backpack::crud.table_max_reached', ['max' => $max])}}'">
+        <table class="table table-bordered table-striped m-b-0" ng-init="field = '#{{ $field['name'] }}'; items = {{ json_encode($items) }}; max = {{$max}}; min = {{$min}}; maxErrorTitle = '{{trans('backpack::crud.table_cant_add', ['entity' => $item_name])}}'; maxErrorMessage = '{{trans('backpack::crud.table_max_reached', ['max' => $max])}}'">
 
             <thead>
                 <tr>